:recycle: 実行中のタスクのストップウォッチを自動で開始するようにした。
