+++ app/src/components/TaskController/index.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
 
 import { useStopwatch } from '../../hooks/useStopwatch/useStopwatch'
 import { Task } from '../../models/task'
@@ -7,15 +7,18 @@ import { Buttons } from './Buttons'
 import { DisplayTime } from './DisplayTime'
 import { Inbox } from './Inbox'
 
-
-
-
-export const TaskController = ({ ongoingTask }: { ongoingTask: Task }) => {
+export const TaskController = ({ ongoingTask }: { ongoingTask: Task | undefined }) => {
   const { isRunning, elapseTime, startStopwatch, stopStopwatch } =
     useStopwatch()
   const hour = Math.floor((elapseTime / (60 * 60)) % 24)
   const minute = Math.floor((elapseTime / 60) % 60)
   const second = elapseTime % 60
+
+  useEffect(() => {
+    if (ongoingTask) {
+      startStopwatch(ongoingTask.start)
+    }
+  }, [ongoingTask, startStopwatch])
 
   const [taskName, setTaskName] = useState<string>(ongoingTask?.name ?? '')
