+++ app/src/components/TaskManager/index.tsx
@@ -2,7 +2,7 @@ import { Task } from '../../models/task'
 import { Tasks } from '../Tasks'
 
 type TasksProps = {
-  tasks: Task[]
+  tasks: Task[] | undefined
 }
 
 export const TaskManager = ({ tasks }: TasksProps) => {
